Added .gitignore to Tychaia.Content.
Summary:
Because of the way `arc diff` works, it'll detect any changed files inside
submodules as uncommitted changes, which basically means that `arc diff`
is hard to use.

This adds a .gitignore to Tychaia.Content to reduce the likely chance that
submodule will have changes in it when trying to use `arc diff`.

If submodules do have changes in them, you need to run `git reset --hard HEAD`
from **within the submodule folder**.  If you do it outside the submodule
folder, well, you'll reset all the changes you were trying to diff on..

Test Plan:
Apply it and open Git Extensions of the submodule.  At worst, the only changes
that will be present in Tychaia.Content is for `Tychaia.Content.proj`.  All
of the "untracked files" inside `compiled\bin` should disappear.

Reviewers: Sabre070

Reviewed By: Sabre070

Differential Revision: http://phab.tychaia.com/D3