Fixing the seed not propagating to inputs when setting it on a layer.
Summary:
Currently the logic just sets the Seed property of the layer before
it called GenerateData on it, however this doesn't actually set the
Seed values of the input layers (nor their input layers and so forth).

This means that effectively the seed value is useless when specified
in pretty much everywhere (in-game, Make Me a World, etc.)

Test Plan:
Run the Make Me a World website locally and give different seeds to
the site.  After applying this patch, the different seeds should generate
different results (just remember to clear the `App_Cache` folder if
comparing results of the same seed between versions).

Reviewers: Sabre070

Reviewed By: Sabre070

Differential Revision: http://code.redpointsoftware.com.au/D4