Implemented initial build generation system.
Summary:
This is the initial implementation of a build generation system.  The idea is that it will generate the project files for MonoDevelop (or the build server) based on the platform the user wants to target.

This is done by defining an MSBuild task which reads in and runs an XSLT document over a bunch of definition documents (located at `Build\Projects`).  From there it reads the information and writes out the project files in the correct directories.

Once we have project definitions for everything (that work for both Windows and Linux), the intention is to delete all of the `.csproj` and `.sln` files in the repository, with instructions for users to build the `Main.proj` project with MSBuild like so:

```
msbuild Build\Main.proj /p:TargetPlatform=Windows
```

Alternatively if you are using MonoDevelop:

```
xbuild Build\Main.proj /p:TargetPlatform=Windows
```

Test Plan:
Run the `Build\Main.proj` script as indicated above.  It should generate a solution file called `Tychaia.Windows.sln` in the root of the project.  Open it up and it should contain 4 projects (the projects this diff defines).

In addition, make sure that Tychaia.Windows references Tychaia.Game.Windows correctly, as this indicates it is correctly resolving the right platform-specific projects.

Reviewers: Sabre070

Reviewed By: Sabre070

Maniphest Tasks: T33

Differential Revision: http://code.redpointsoftware.com.au/D10