Finishing conversion of projects to new build system.
Summary:
This diff finishes converting all of the projects to use the new build system, including:

  * Definition files for all of our projects and external libraries.
  * Support in the project generator for NuGet and multi-platform libraries.

It also updates the MonoGame and Voxeliq submodules.  MonoGame is updated to use the upstream repository as we now no longer need patches for it to work correctly with Tychaia (the workarounds we were using are now properly fixed in upstream).

This diff *does not* remove the old solution and project files.  That will be done in a second diff if this one passes code review.

Test Plan: Use the new build system to generate solution and project files for your platform (Windows or Linux).  The build should complete with 0 errors across all projects.  You should be able to run MakeMeAWorld, Tychaia.Website and the main Tychaia game without issue.

Reviewers: Sabre070

Reviewed By: Sabre070

Differential Revision: http://code.redpointsoftware.com.au/D11