Migrated test system to xUnit instead of NUnit
Summary:
Currently NUnit has a few bugs which means that it can't properly be
used under Linux.  This migrates our test projects over to use xUnit
which should work under both platforms.  In addition, xUnit allows test
execution to be directly part of the build, so when we build the
project, the tests are automatically run.

Test Plan:
Apply the diff and regenerate the projects with the new build system.
Once this is done, build the solution with `msbuild` or `xbuild` from
the command line (building under MonoDevelop does not run the tests).

Reviewers: Sabre070

CC:

Maniphest Tasks: T4