Implemented threaded tasking region and updated xUnit tests.
Summary: This implements the logic around task regions, where processing the region causes each of the individual cells in the octree to be processed on a thread.  This will be used to implement chunk processing and rendering in the main game.

Test Plan: Regenerate the projects and run `xbuild`.  The tests should all pass.

Reviewers: Sabre070

Reviewed By: Sabre070

Maniphest Tasks: T4

Differential Revision: http://code.redpointsoftware.com.au/D14