Improvements to the game's title screen.
Summary:
This adds a parallex background to the game's title screen and re-arranges some of the UI as well.

It also contains a fix for the `RenderingBuffers` class, which didn't clear the surface after they were first initialized, nor did it resize the buffer when the user resized the window.  Thus the surface could contain garbage graphics content and cause artifacts when rendered.

Test Plan: Regenerate the projects and run `xbuild`.  You should be able to run the game and see the parallex background.  There shouldn't be any graphical artifacts (such as messed up graphics).

Reviewers: Sabre070

Reviewed By: Sabre070

Maniphest Tasks: T7

Differential Revision: http://code.redpointsoftware.com.au/D15