Add notification when the game can't start due to lack of 3D hardware acceleration
Summary: This adds a notification at game startup if the game fails to initialize OpenGL.  This is mainly about making it more friendly when trying to start the game (as opposed to it just crashing).

Test Plan: On computers with OpenGL hardware acceleration, the game should run fine.  On those without (such as a Windows VM) it should display the notification accordingly.

Reviewers: Sabre070

Reviewed By: Sabre070

Maniphest Tasks: T16

Differential Revision: http://code.redpointsoftware.com.au/D18