Added Arcanist workflow for checking environment conditions.
Summary: Added a new workflow for Arcanist that can be invoked as `arc check-env`.  This currently just checks for the issue with Mono where dmcs / gmcs aren't setup as batch scripts but in future we can extend it to check various other things.

Test Plan: Run `arc check-env`.  It should tell you all of the environment conditions pass.  If you rename `dmcs.bat` in your Mono bin directory it should then fail with an appropriate message.

Reviewers: Sabre070

Reviewed By: Sabre070

Differential Revision: http://code.redpointsoftware.com.au/D31