Fixing issue where unit test engine no longer worked on Linux after D30.
Summary: D30 broke the unit test engine on Linux.  This fixes it so that the Mono runtime engine never prefixes the `xbuild` command, since on both platforms `xbuild` is a shell script that already wraps the call in whatever runtime is appropriate.

Test Plan: Make sure the unit test engine works on Windows with `arc unit --everything`.

Reviewers: Sabre070

Differential Revision: http://code.redpointsoftware.com.au/D34