Added cscjvm to libraries.
Summary: This is adding cscjvm support to Tychaia.  Currently cscjvm is highly unstable and most assemblies don't compile (where they do, they aren't valid code).  I'll set up a build server entry to build with cscjvm so we can monitor what parts of the codebase are breaking.  Whether or not cscjvm is really a feasible option long term at this stage I'm not too sure; there's still quite a lot of areas of the C# language that are complex (such as yielding and async support).

Test Plan: Make sure all of the unit test are still passing for Linux / Windows.  The Java platform target at this point in time is definitely broken.

Reviewers: Sabre070

Reviewed By: Sabre070

Maniphest Tasks: T40

Differential Revision: http://code.redpointsoftware.com.au/D33