Implementation of build server cache on latest infrastructure.
Summary: D9 no longer applies cleanly on top of the latest version of the repository due to some restructuring we did during the last week.  This is a reimplementation of the same functionality, but it also includes unit tests as well.

Test Plan: Run the Tychaia website and make sure the download page still works (it should render the links given the build server is online).  Also run the unit tests to make sure everything is working on Windows.

Reviewers: Sabre070

Reviewed By: Sabre070

Differential Revision: http://code.redpointsoftware.com.au/D36