Updating unit test engine to ignore changes to .dll and .mdb files.
Summary: The unit test engine was previously picking up the changed .dll / .mdb files as a reason to run the unit tests.  This meant that regardless of whether or not you changed code files in the website, it would always trigger the unit tests for it since the files are always changed after a build.

Test Plan: Rebuild and regenerate, then run the unit tests.  Given that the only changed files are the DLLs / MDB files for the website, the unit tests for the website should not trigger.

Reviewers: Sabre070

Reviewed By: Sabre070

Maniphest Tasks: T13

Differential Revision: http://code.redpointsoftware.com.au/D39