Added more unit tests for the lightspeed compiler.
Summary:
This adds more unit tests for the lightspeed compiler, which seems to indicate that after the changes in D40, the existing implementation is working (and resulting in identical output to the runtime layer).

Depends on D40.

Test Plan: Run the unit tests.  It would also be ideal to do some manual verification (maybe hook up `ProceduralGenTracing` to compile the layers) to verify that the output is actually identical and that these unit tests aren't lying.

Reviewers: Sabre070

Reviewed By: Sabre070

Maniphest Tasks: T13

Differential Revision: http://code.redpointsoftware.com.au/D41