Replaced terrible global class for rendering location with IoC interface.
Summary:
Currently there is a class called `TemporaryCrapBecauseIDidNotDesignThingsVeryWell`.  This was prior to the introduction of IoC and was used so that the rendering classes (statically implemented) could find the location that the rendering needed to happen at (since the rendering logic implicitly generated the data for the render).

This abstracts the location functionality out into an IoC interface and class where the form implements the interface to provide the location straight from the form element values.

Test Plan: Make sure rendering still works in the world generation editor.

Reviewers: Sabre070

Reviewed By: Sabre070

Differential Revision: http://code.redpointsoftware.com.au/D42