Added visitor to clean up temporary __c* variables.
Summary: Fixes T45 by adding a visitor to inline all of the __c* variables.  One of the side effects of this is that afterwards we can see that the `__cx` offsets are being applied to everything redundantly; the loop starts at `-6, -6, -6` for the configuration generated by `ProceduralGenCompiler`, but then everywhere that `i`, `j` and `k` are referenced it adds `6` back on again.  Thus we should probably also look into whether or not we can make these offsets cancel out in the resulting code as well and this will help clear up the reason why the current tests fail with "Array index out of bounds" (once such a visitor is applied everything will be relative to `0, 0, 0` in the arrays, which should make spotting invalid array indexes much easier).

Test Plan: Run `ProceduralGenCompiler`; there should be no instances of `__cx`, `__cy`, `__cz`, `__cwidth`, `__cheight`, `__cdepth` in the output and the layer should compile successfully (but fail when trying to get data with "Array index out of bounds").

Reviewers: Sabre070

Reviewed By: Sabre070

Maniphest Tasks: T45

Differential Revision: http://code.redpointsoftware.com.au/D45