Updated force-reset workflow to also fetch all submodules.
Summary:
The `arc force-reset` command now supports a `--fetch` argument, which can be used to fetch all of the latest commits in the submodules.  It's an optional argument since it takes a while to perform regardless of whether there are any new changes (since it has to contact each of the git repositories to check).

Note that this only fetches the new commits, if you're doing `arc patch` you'll still need to use `git checkout` in submodules to update them to the correct version while https://secure.phabricator.com/T3440 is still outstanding (the reason we can't just do `git checkout master` is that some submodules are on different branches).

Test Plan: Check the latest commit in, for example, `Libraries/MonoGame`.  Now run `arc force-reset --fetch`.  All of the submodules should have new commits pulled down and you should be able to see new commits in `Libraries/MonoGame`.

Reviewers: Sabre070

Reviewed By: Sabre070

Differential Revision: http://code.redpointsoftware.com.au/D54