Implemented UI framework for asset manager.
Summary: This implements a UI framework for use in the asset manager (but the flexible skinning system means that we can also use it for in-game UI as well).  It updates the asset manager to display the status and the list of assets using the UI elements, and the "Mark Dirty" operation is now a button in the asset manager.

Test Plan: Test Tychaia running with the asset manager (as before, use `Tychaia.exe --asset-manager`) and ensure that the UI behaves as expected.  Clicking the "Mark Dirty" button should cause the assets to be marked as dirty, and a list of assets should be displayed (generally) in the center of the asset manager (in the third section).  The last button and the last label do nothing (and are only for testing that specifying the sizes of each child works correctly).

Reviewers: Sabre070

Reviewed By: Sabre070

Maniphest Tasks: T51

Differential Revision: http://code.redpointsoftware.com.au/D55