Implemented support for local asset loading and asset editors.
Summary: This adds support for loading in assets from the local directory when the asset manager is started on it's own, as well support for asset editors.  Depends on D57.

Test Plan: Start the asset manager on it's own.  When selecting text assets, it should display the text value in the editor area (there are no textbox controls yet).

Reviewers: Sabre070

Maniphest Tasks: T51

Differential Revision: http://code.redpointsoftware.com.au/D58