Implemented textbox control and local and network saving of text assets.
Summary: Implemented a textbox control and finished off the implementation for saving and baking assets in the asset manager.

Test Plan:
Run the game with the asset manager.  Editing a text asset should update the content in the game in (almost) real-time.  If you observe the files at `Tychaia/bin/Debug/Content/assets/language/`, the asset files should only change to the new content when the `Bake` button is pushed.

The same goes for running just the asset manager by itself, although in this case obviously nothing will happen as you change the text; only baking will cause the changes to be saved to disk.

Reviewers: Sabre070

Maniphest Tasks: T51

Differential Revision: http://code.redpointsoftware.com.au/D59