Support for in-game console (press '~' key to open).
Summary:
While working on improvements to the chunk rendering / caching system (to make it work faster), I needed a way to change the camera focus at runtime.  The best way to do this was to implement a console that can be opened with the '~' key so that things can be changed at runtime.

Part of this infrastructure (the core bits like the console interface `IConsole` and implementation `DefaultConsole`) are pushed to the Protogame repository.  This revision only covers the changes in Tychaia.

This adds a `CameraCommand` which allows you to enable / disable rotation, and set the distance from the focal point for the camera.

Test Plan:
Build and run the game in Debug mode.  You should be able to open the console with the '~' tilde key.  If you type 'help', you should see a list of commands.  Try using the 'camera' command to enable / disable rotation, as well as change the camera focus distance.

If you build and run in Release mode, you shouldn't be able to open the console with the '~' tilde key.

Reviewers: Sabre070, jrhodes

Reviewed By: jrhodes

Maniphest Tasks: T97

Differential Revision: http://code.redpointsoftware.com.au/D61