Implemented persistent storage and settings.
Summary: This implements persistent storage that we can use to store game settings.  It also implements a command to disable the per-method profiler (the green graph when running the game in debug mode) as this causes some performance issues when playing the game.

Test Plan: Run the game in debug mode; the green graph should display in the profiler area.  Open the console with `~` and type `profiler per-method disable` followed by `exit`.  Run the game again and the green graph should no longer display.  Verify that in `%appdata%\.tychaia\settings` that there is JSON with the `PerMethodProfiling` set to false.  Run the game again and open the console with `~`, type `profiler per-method enable`.  Without exiting, verify that the changes were saved into the settings file after typing the command.  When you restart the game, the green graph should display again.

Reviewers: Sabre070, jrhodes

Reviewed By: jrhodes

Differential Revision: http://code.redpointsoftware.com.au/D69