Implemented base game entities and inventory class.
Summary: This implements the base game entities and the inventory class in the Tychaia.Game assembly.  It also adds a set of tests for the inventory to verify the behaviour.

Test Plan: Make sure that the implementation of the inventory matches the requirements for the in-game inventory (i.e. it meets the needs of what we need to do with inventories in the game).

Reviewers: Sabre070

Reviewed By: Sabre070

Maniphest Tasks: T104

Differential Revision: http://code.redpointsoftware.com.au/D70