Json.NET upgrade, NoProfile attribute and fix persistent storage not creating directories.
Summary:
This updates Protobuild to use Json.NET, which greatly improves the performance and memory usage when loading and saving assets.

It also adds a NoProfile attribute to Tychaia.Globals, so that services can be marked with it.  When a service is marked with the NoProfile attribute, the per-method profiler will never intercept this service.  This ensures the profiler does not create too much overhead on highly accessed services (such as scaling services where there is no actual behaviour; only properties).

In addition, there is also a fix for persistent storage not creating the `.tychaia` directory when needed.

Test Plan: N/A

Reviewers: Sabre070

Reviewed By: Sabre070

Differential Revision: http://code.redpointsoftware.com.au/D72