Fixing issue with position octree scale translation with negative values.
Summary: This fixes the issue where translation of negative values from in-game coordinates to the position octree coordinates would result in the wrong values.  It seperates the logic out into it's own class and adds a unit test to ensure that this does not break in the future.

Test Plan: Run the game, you should be able to correctly walk into the negative region now.

Reviewers: Sabre070

Reviewed By: Sabre070

Differential Revision: http://code.redpointsoftware.com.au/D71