Fix bug in AlgorithmReduceWaterDepth
Summary:
This fixes a bug in AlgorithmReduceWaterDepth, where it would incorrectly access 3D array indexes even though it is a 2D layer.  It also removes a `try { } catch { }` around image generation so that crashes aren't silenced when the world editor is being run.

Update C# unit test to allow coverage to be disabled.

Test Plan: The layer should not crash.

Reviewers: Sabre070, buildbot

Reviewed By: Sabre070

CC: buildbot

Differential Revision: http://code.redpointsoftware.com.au/D79