Increase performance of code coverage.
Summary: This increases the performance of the code coverage option by using a version of cstools that sets fields instead of calls a function to work out what lines of code have been called.  In addition, there's been some changes in the Arcanist integration that reduce the amount of XML that needs to be parsed, and it also adds support for the `unit.csharp.coverage.match` regex option so that the number of assemblies that are instrumented can be reduced as well.

Test Plan: Run `arc unit --coverage`.  Things should be a lot faster.

Reviewers: Sabre070

Reviewed By: Sabre070

CC: buildbot

Differential Revision: http://code.redpointsoftware.com.au/D80