Add basic UI for a status bar.
Summary: This adds a really basic UI for a status bar.  There's a vertical container on each side of the screen, currently set to 0 pixels wide; these will enable the inventory pop-out like you see in Diablo II.

Test Plan: Run the game, you should see a red rectangle at the bottom of the screen.  If you resize the window, it should remain the same size and in the horizontal centre of the screen.

Reviewers: Sabre070

Reviewed By: Sabre070

CC: buildbot

Maniphest Tasks: T106

Differential Revision: http://code.redpointsoftware.com.au/D81