Add option to toggle profiler visibility in Debug mode.
Summary: This adds two new commands: `profiler visibility hide` and `profiler visibility show`.  These toggle whether profiling information is shown in the top-left of the screen when the game is running in Debug mode.  In my particular case I needed some way of hiding it so I could see the sidebars and game content properly.

Test Plan: Run the game and use the specified commands to toggle whether the profiler information shows up.

Reviewers: Sabre070

Reviewed By: Sabre070

CC: buildbot

Differential Revision: http://code.redpointsoftware.com.au/D82