Upgrade Tychaia to use Protogame event system and change over PlayerEntity.
Summary: This updates Tychaia to use the new event engine in Protogame.  It also updates PlayerEntity so that movement is driven by bound keyboard events.

Test Plan: Apply the patch and set a breakpoint on the `MoveInDirection`.  As this method is called from the `MoveForwardAction`, verifying that this method is called ensures that the event system is working (in addition to the fact the old movement code has been removed).

Reviewers: Sabre070, buildbot

Reviewed By: Sabre070

CC: buildbot

Maniphest Tasks: T111

Differential Revision: http://code.redpointsoftware.com.au/D87