Check for long frames in the profiler.
Summary: This adds a check to the profiler to output the state of the frame when the frame takes longer than 16ms (i.e. the FPS drops below 60).  This allows us to inspect long frames because we can then see exactly what took the most time in the frame.

Test Plan: Run the game, you should see long frames having their stats outputted to the console.

Reviewers: Sabre070

Reviewed By: Sabre070

CC: buildbot

Differential Revision: http://code.redpointsoftware.com.au/D89