Update inventory UI to use new events system.
Summary: This updates the inventory UI to use the new events system.  In addition, the sidebars have been remapped to keys `C` and `I` (for character and inventory respectively).

Test Plan: Run the game and press `C` and `I` to open the sidebars.  If you open the console with `~`, then as you type, pressing `C` and `I` shouldn't cause the sidebars to open (and the console should consume the characters).

Reviewers: Sabre070

Reviewed By: Sabre070

CC: buildbot

Maniphest Tasks: T111

Differential Revision: http://code.redpointsoftware.com.au/D90