Implemented inventory manager and pulling through inventory to UI.
Summary: This implements the inventory manager UI and hooks it up to the player's inventory.  You can then give the player items with the console command `give something` and it will appear in the inventory UI (toggled with the `R` key in this diff).

Test Plan: Use the console command `give <anything>` to add items to the player's inventory.  Verify that they appear in the inventory UI (press `R`).

Reviewers: Sabre070

Reviewed By: Sabre070

CC: buildbot

Maniphest Tasks: T106

Differential Revision: http://code.redpointsoftware.com.au/D92