Enhance profiler to monitor individual frame timings and RAM usage.
Summary:
This updates the profiler so that it monitorings individual frame timings and displays a benchmark for 60FPS as a red line in the profiler display.  In addition, it shows the RAM usage in purple.

If you run the game, this helps identify when the GC is kicking in due to excessive memory usage.

Test Plan: Run the game and observe the profiler information in the top-left of the screen.

Reviewers: Sabre070

Reviewed By: Sabre070

CC: buildbot

Maniphest Tasks: T112

Differential Revision: http://code.redpointsoftware.com.au/D94